user/recover_password/requirements.txt
falcon==1.4.1
gunicorn==19.9.0
pymongo==3.7.2
pyjwt==1.7.1
falcon-cors==1.1.7
redis==2.10.6
rq==0.12.0
passlib==1.7.1
sendgrid==5.6.0